IntroductionToConstructors
In this project, we will discuss the use of constructor methods in Java. Constructors are special methods that are called when an object of a class is created. Constructors are used to assign initial values to objects and have the same name as the class name. In this example we will create a Person class.